Use flash() instead of passing error as context.
In some cases I wanted to pass an error message with a redirect, but
this was complicated because you have to pass variables to url_for as
part of the query string. Then I'd have to check for errors within
request.args and that's needlessly complicated. Just going to flash the
error messages instead.